backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
ecdsa==0.18.0
Flask==2.3.2
Flask-Cors==4.0.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pyasn1==0.5.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
